W200: the draw-call live-value idiom — +4 MATCH (6523 -> 6527, 79.5%), one window cleared
All four MATCHes are in mt_evtMenu and all four come from the idiom W199 discovered
there, now fully characterised.  0 MATCHes lost, 0 new over-window, 0 compile errors.

THE IDIOM, in its final form:

  * ONE real scaled value per switch case, NAMED and retained across that case's PAIR
    of draw calls — not recomputed, and not a function-wide coordinate reused.
  * ASCENDING switch case bodies.
  * An ADDRESSED two- or four-float stack aggregate assigned as a whole `RwV2d`/`RwV4d`,
    which is what reproduces retail's load-load / store-store order.  **This is the new
    half and it is what converted the near-misses**: 0037dfc0 (nd7) and 00385260 (nd7)
    had both been standing `confirmed floors` for several waves and both MATCHed the
    moment their two float labels became one addressed aggregate copied by assignment.
  * Typed absolute string objects for low-memory literals, GP-RELATIVE objects for
    strings inside GP range.

  mt_evtMenu 00384520  nd18  -> MATCH   (nd379 two waves ago)
  mt_evtMenu 00380a30  nd556 -> MATCH
  mt_evtMenu 0037dfc0  nd7   -> MATCH
  mt_evtMenu 00385260  nd7   -> MATCH
  mt_evtMenu 0037f650  nd515 -> nd483   one genuinely scoped retained scaled-Y in each
                                        of eight draw-pair cases, replacing a
                                        function-wide reused coordinate

Also kept:
  mdlFile 00320de0     404/400 OVER-WINDOW nd257 -> 396/400 nd17.  FUN_00320de0's
                       callback contract is u32, not u64; the u64 spelling was emitting
                       a spurious `dsll32`/`dsra32` pair.  Window cleared.
  op_wait 0026eed0     nd6307 -> nd6245, -40 bytes.  A corrected typed
                       OpWaitWork/OpWaitPart layout, then the real oscillating scale
                       computed once and retained across both quad-axis consumers
                       instead of calling the sine function twice.
  fclShopMisc 003f6f20 nd779 -> nd753, -32 bytes, on one genuinely computed half-alpha
                       retained across the paired empty-slot draws.
  btlVoice 002ee640    nd458 -> nd444: retail's DESCENDING two-case body layout, typed
                       absolute objects for both low-memory string literals, and the
                       parameter contract corrected from `u64*` plus an invented action
                       alias to a real `BtlAction*`.
  k_clump 001a7710     byte-neutral source cleanup: an explicit goto guard became the
                       equivalent null early return.

WHERE THE IDIOM DOES **NOT** APPLY, which is as useful as where it does.  It was tested
on fourteen further draw routines and rejected on all of them: mt_evtMenu 0037ad30 and
0037b350; fclShopMisc 003f5830, 003f7390, 003f7890, 003f7730; _h_camp_status 0012b860
and 0012bfb0 (nd322 -> 758, and a corrected callback ABI separately went over-window at
nd694).  On the three game_support_late siblings the retained half-width/half-height was
kept as honest source but is **byte-identical — MWCC already CSEs the repeated
divisions**, which is the precise boundary: the idiom only pays where retail's value
crosses a CALL, because that is where b210 cannot CSE it.

THE NEAR-MISS BAND IS NOW RIGOROUSLY CLOSED, not merely labelled.  Thirty-odd rows were
decoded end to end and their residuals written out instruction by instruction.  The
recurring families, with representative sites:

  * commutative operand orientation — `addu v1,v1,v0` vs retail `addu v1,v0,v1`
    (itfPanel 003ad640 off32, frTagTable 003af660 off36), and commutative `mul.s`
    operand order (bp_dialog 0025c220 off1072, sfl_count 0025b690, btlCamera 002b17a0,
    mt_sceneFunc 003bbb90 off180).  Syntactic swaps either regress or are byte-neutral.
  * cyclic register colouring — itfPanel 003ae160 and 003ae260 share the *identical*
    seven-word residual at offsets 80/156/172/184/196/204/216 (a2/a3, v1/a2, a3/v1);
    itfPanel 003ada00 five words s1/s2; mdlFile 0031d7e0 param_3/param_4 as s3/s2 vs
    retail s2/s3; mdlFile 00320380 a consistent s0/s1 rotation; mdlManager 003197c0
    a2/a3; mdlFile 00332470 exactly three dependent rows (`andi` v1 vs a2, its `sltiu`,
    its `sll`); datUnit 002ff3e0; k_fldrc 001eba80 (switch discriminator allocation).
  * pre-JAL argument setup order — mt_sceneFunc 003bbed0 (offsets 200/204),
    btlUnit 00289650 (now permanently annotated in source), btlAction 00297480,
    btlFormation 002bf9b0, mdlFile 00332880, k_fldCamera 001d6b10, sfl_card 00258b40,
    y_runtime 00449fe0 and 0044a110 (twins, three words each), 00457470.
  * `addiu` vs retail `daddiu` when materializing a small constant — btlBoss 002f88c0
    (the whole residual), btlCamera 002b17a0.  Explicit 64-bit literals and s64/u64
    counters were tried; they either do nothing or regress badly (nd3 -> nd88).
  * _h_camp_main 00136750 is now closed exactly: all 49 body instructions match and the
    ONLY difference is our 0x40 frame against retail's 0x50.  Growing it would require
    an unused local, which is banned; this is the correct permanent outcome.
  * mt_evtMenu 0036f900, after seven reconstructions across the project's history, is
    finally pinned: of 80 instructions exactly ONE differs — at +0xA8 ours is
    `beqz $s7, 0x36fa00` where retail targets 0x36f9cc, the unconditional branch at
    +0xCC.  Nesting the loop under the callback guard moved the target to 0x36fa0c.
  * mt_sceneFunc 003bbb90 carries a genuine TYPE finding rather than a floor: at off32
    we emit `ld v0,8(base)` where retail emits `lwc1 f0,8(base)`, and at off40 `sd`
    where retail has `swc1`.  We copy a float field as an integer doubleword.  The
    scalar field-copy probe regressed, but given W200's result on 0037dfc0/00385260
    the answer is very likely an addressed float aggregate assignment.

Two files are confirmed MINED OUT as template sources: btlCamera (zero movement from
eight transfers, twice) and the itfMesManager / itfPanel / frFont group (zero from
eight, every rewrite regressing, including the levers that originally produced fifteen
MATCHes there).

+++ src/Graphics/Model/mdlFile.c
@@ -789,7 +789,7 @@ u32 FUN_00320880(u64 param_1,int *param_2,int param_3,u64 param_4);
 u32 FUN_00320ba0(u32 param_1,u32 *param_2);
 u32 FUN_00320cf0(u32 param_1,u32 param_2);
 u64 FUN_00320da0(u64 param_1,u64 param_2);
-u64 FUN_00320de0(u64 param_1,u32 *param_2);
+u32 FUN_00320de0(u32 param_1,u32 *param_2);
 u16 FUN_00320f70(u64 param_1,u32 param_2);
 u32 FUN_00320fb0(u16 param_1);
 void FUN_00321010(int param_1);
@@ -5193,7 +5193,7 @@ u64 FUN_00320da0(u64 param_1,u64 param_2)
 // FUN_00320DE0 NONMATCHING
 
 
-u64 FUN_00320de0(u64 param_1,u32 *param_2)
+u32 FUN_00320de0(u32 param_1,u32 *param_2)
 {
   s32 materialCount;
   s32 arrayCount;
